Adding init-bootsplash for configuring a bootsplash from react-native command line with arguments
